part 4

+++ part4/Bloglist/controllers/blogs.js
@@ -1,22 +1,22 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
 const User = require("../models/user")
+const jwt = require("jsonwebtoken")
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog
-    .find({}).populate("user")
+    .find({}).populate("user", { username: 1, name: 1 })
 
   response.json(blogs.map(blog => blog.toJSON()))
-
-  // Blog
-  //   .find({})
-  //   .then(blogs => {
-  //     response.json(blogs)
-  //   })
-  //   .catch(error => error.message)
 })
 
 blogsRouter.post("/", async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!(request.token || decodedToken.id)) {
+    return response.status(401).json({ error: "token missing or invalid" })
+  }
 
   const user = await User.findById(request.body.userId)
 
@@ -40,18 +40,39 @@ blogsRouter.post("/", async (request, response) => {
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
   response.status(201).json(savedBlog.toJSON())
-  //   .save()
-  //   .then(result => {
-  //     response.status(201).json(result)
-  //   })
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!(request.token || decodedToken.id)) {
+    return response.status(401).json({ error: "token missing or invalid" })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
 
 blogsRouter.put("/:id", async (request, response) => {
+  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  // if (!(request.token || decodedToken.id)) {
+  //   return response.status(401).json({ error: "token missing or invalid" })
+  // }
+
+  // const blog = await Blog.findById(request.params.id)
+
+  // if (request.user.toString() !== decodedToken.id) {
+  //   return response.status(401).json({ error: "token missing or invalid" })
+  // }
+
   const result = await Blog.findByIdAndUpdate(request.params.id, request.body, { new: true })
   response.status(200).json(result)
 })
